// tools/code/xml-to-json-converter/src/types.ts
export type XmlToJsonOptionsConfig = {
  compact: boolean
  ignoreDeclaration: boolean
  ignoreInstruction: boolean
  ignoreAttributes: boolean
  ignoreText: boolean
  ignoreCdata: boolean
  ignoreDoctype: boolean
  ignoreComment: boolean
  trim: boolean
  nativeType: boolean
  alwaysArray: boolean
  alwaysChildren: boolean
}